fix segfault

+++ src/services/count.h
@@ -5,7 +5,7 @@
 class ServiceCount : public TraverseCombinePST {
 public:
     typedef struct Cell input; // Array of Cells
-    typedef uint64_t output;   // Array of counts
+    typedef unsigned int output;   // Array of counts
     explicit ServiceCount(PST pst)
         : TraverseCombinePST(pst,PST_COUNT,MAX_CELLS*sizeof(input),MAX_CELLS*sizeof(output),"Count") {}
 protected:
